% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{image.CornerDetectionF9-package}
\alias{image.CornerDetectionF9-package}
\title{Find Corners in Digital Images with FAST-9.}
\description{
An implementation of the "FAST-9" corner detection algorithm explained at <http://www.edwardrosten.com/work/fast.html>. 
The package allows to detect corners in digital images.
}
\seealso{
\link{image_detect_corners}
}
